# design/verity/contracts/openapi/verity-tools.openapi.yaml
openapi: 3.0.3
info:
  title: Hybrid Retrieval Workbench tool contracts
  version: 1.0.0
  description: >
    Transport-neutral tool contracts for Hybrid Retrieval Workbench.
    Release packaging must replace the placeholder server URL with the deployed
    static endpoint before creating the AgentCore Gateway OpenAPI target.
    Operation IDs are contract-stable; renaming one is a contract-version change.
    AgentCore Gateway does not expose the bare operationId - it prefixes the target
    name as ${targetName}___${operationId}, so target verity-openapi-tools exposes
    search_evidence as verity-openapi-tools___search_evidence. HTTP and stdio MCP use
    the bare operationId. See
    https://docs.aws.amazon.com/bedrock-agentcore/latest/devguide/gateway-tool-naming.html


    Response schemas are widened to the inspectable-output table in spec section 13.
    Every change in that pass is additive - new optional properties, new components,
    and relaxed `required` lists - so contract_version remains 1.0.0. No field, path,
    or operationId was removed or renamed. Two shapes changed meaning and are called
    out in their own descriptions: TraverseResponse.nodes is now an object rather than
    a bare string, and SynthesizeResponse split away from AnswerResponse.


    Ground truth for every example in this document is fixtures/generate.py and the
    fixtures it emits. Numbers here are copied from fixtures/runs/RUN-7000.json and
...
servers:
...
paths:
  /v1/tools/decompose:
    post:
      operationId: decompose_question
      summary: Decompose an incident-evidence question into IDs, filters, subquestions, and retrieval steps.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecomposeRequest'
      responses:
        '200':
          description: Deterministic plan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecomposeResponse'
  /v1/search:
    post:
      operationId: search_evidence
      summary: Run Aurora PostgreSQL hybrid retrieval and persist the candidate receipt.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Candidate receipt - run ID, applied controls, candidate count, arm diagnostics, candidates.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
  /v1/tools/traverse:
    post:
      operationId: follow_evidence_links
      summary: Traverse ACL-visible canonical and governed inferred evidence relationships.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TraverseRequest'
      responses:
        '200':
          description: Relationship paths, per-node depth, edge origin and confidence, and the ACL decision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraverseResponse'
  /v1/tools/compare:
    post:
      operationId: compare_sources
      summary: Compare selected evidence and return confirmed, ruled-out, affected, and unaffected verdicts.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompareRequest'
      responses:
        '200':
          description: Source comparison.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompareResponse'
  /v1/tools/explain-ranking:
    post:
      operationId: explain_ranking
      summary: Human diagnostic surface. Explains one candidate's arm positions, RRF and rerank. NOT an agent step.
      description: >
        The workbench calls this when a participant expands a candidate row. An agent must
        not select it. There is no action an agent can take on "CHG-1000 was text_position
        2" - no branch in the loop reads it, and scripting a planner to call a debugger it
        cannot act on is how a fixed storyboard gets mistaken for a policy. The tools an
        agent actually selects are decompose_question, search_evidence (once per
        subquestion, twice on escalation), follow_evidence_links and compare_sources, with
        answer_with_citations as the orchestration entry point.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExplainRankingRequest'
      responses:
        '200':
          description: Ranking explanation - positions, raw arm diagnostics, per-arm RRF contribution, rerank delta, and the controls that produced the score.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExplainRankingResponse'
  /v1/tools/synthesize:
    post:
      operationId: synthesize_cited_answer
      summary: Server-side citation binding and receipt persistence. NOT an agent step.
      description: >
        Binds each claim to an exact document and chunk version, runs
        proof.validate_answer_citations, and persists the receipt. An agent must not select
        it: answer_with_citations invokes it server-side at the end of the loop, and it is
        exposed as a tool only because the contract has to be portable across HTTP, stdio
        MCP and Gateway - not because a planner would choose it. Citation validation
        failure is a hard stop, never a warning: one bounded repair attempt, and if the
        repaired answer still fails the run ends `failed` and no answer text is returned.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SynthesizeRequest'
      responses:
        '200':
          description: Cited answer, per-citation claims, and synthesis mode.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SynthesizeResponse'
  /v1/agent/answer:
    post:
      operationId: answer_with_citations
      summary: Run the complete inspectable Hybrid Retrieval Workbench agent path.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerRequest'
      responses:
        '200':
          description: >
            Plan, retrieval receipt, graph, and cited answer. Also returned when every arm
            was empty: candidate_count 0 with synthesis_mode `no_evidence` is a successful
            query that found nothing, not an error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnswerResponse'
        '422':
          description: >
            Citation validation failed and the bounded repair attempt did not fix it. No
            answer text is returned under any status code when a citation does not
            validate. The body is the failing rows from proof.validate_answer_citations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CitationValidationError'
  /v1/agent-runs/{agent_run_id}:
    get:
      operationId: get_agent_run
      summary: Read a persisted agent run - subquestions, coverage, escalations, budget spend.
      description: >
        Workbench read surface, NOT an agent tool. An agent never calls this; it is how the
        Hybrid Retrieval Workbench UI and a replaying participant fetch a completed proof.agent_runs row with
        its subquestions, retrieval attempts, coverage decisions and escalations. Every
        value it returns was persisted during the run, so a replay renders identically to
        the live run.
      parameters:
        - name: agent_run_id
          in: path
          required: true
          description: Agent run identifier, for example ARUN-8000.
          schema:
            type: string
      responses:
        '200':
          description: The complete agent run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRunResponse'
  /v1/agent-runs/{agent_run_id}/coverage:
    get:
      operationId: get_agent_coverage
      summary: Read the per-subquestion coverage decisions for an agent run.
      description: >
        Workbench read surface, NOT an agent tool. Returns just the coverage decisions -
        which subquestions were covered, which kinds were missing, which evidence satisfied
        each kind, and how many attempts it took - without the candidate receipts. This is
        the panel a participant reads next to the SQL rule in proof.evaluate_subquestion_coverage
        to confirm the agent's escalation decision was a predicate and not a judgement.
      parameters:
        - name: agent_run_id
          in: path
          required: true
          description: Agent run identifier, for example ARUN-8000.
          schema:
            type: string
      responses:
        '200':
          description: Coverage decisions for every subquestion in the run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentCoverageResponse'
components:
  schemas:
...
    DecomposeRequest:
      type: object
      required: [question, principal]
      properties:
        question:
          type: string
        principal:
          $ref: '#/components/schemas/Principal'
    DecomposeResponse:
      type: object
      description: >
        The deterministic plan. Identifier tokens are extracted with [A-Z]{2,6}-[0-9]{3,6}
        and classified into exactly two disjoint sets, both of which travel with the
        retrieval request: `id_tokens` resolve to an external_key exactly and drive the
        exact/B-tree half of the text arm; `unresolved_id_tokens` resolve to nothing and
        are the only trigram probe set. `detected_ids` is their union and is retained
        unchanged.
      required: [contract_version, request_id, detected_ids, filters, subquestions, steps]
      properties:
        contract_version:
          type: string
        request_id:
          type: string
        detected_ids:
          type: array
          items:
            type: string
        id_tokens:
          type: array
          description: Detected tokens that match an external_key exactly.
          items:
            type: string
        unresolved_id_tokens:
          type: array
          description: >
            Detected tokens that match nothing exactly. These, and only these, are probed
            with pg_trgm. Decomposing the fuzzy preset yields id_tokens [INC-2000] and
            unresolved_id_tokens [CGH-1000].
          items:
            type: string
        filters:
          $ref: '#/components/schemas/Filters'
        subquestions:
          type: array
          description: >
            TYPE CHANGE, additive within contract_version 1.0.0. Items were previously bare
            strings and are now Subquestion objects carrying subquestion_id, text and
            required_kinds. A pre-widening consumer that read `subquestions[i]` as a string
            must read `subquestions[i].text`. The change is not cosmetic: the plan's
            subquestions used to be emitted and never consumed, and required_kinds is what
            the agent loop branches on - it is the input to the deterministic coverage rule
            that decides whether to escalate. A plan whose subquestions carry no
            required_kinds cannot drive the loop. Only subquestion_id, text and
            required_kinds are populated here; attempts, runs and final_coverage are filled
            in on the agent run, once retrieval has actually happened.
          items:
            $ref: '#/components/schemas/Subquestion'
        steps:
          type: array
          items:
            type: string
    SearchRequest:
      type: object
      required: [query, principal]
      properties:
        query:
          type: string
        principal:
          $ref: '#/components/schemas/Principal'
        filters:
          $ref: '#/components/schemas/Filters'
        controls:
          $ref: '#/components/schemas/SearchControls'
        id_tokens:
          type: array
          description: >
            Optional. Identifier tokens already resolved by decompose_question. Supplying
            them lets the service skip re-resolution; omitting them makes the service
            derive both token sets from `query`.
          items:
            type: string
        unresolved_id_tokens:
          type: array
          description: >
            Optional. The trigram probe set from decompose_question. An empty array means
            every identifier in the query resolved exactly and the fuzzy arm is legitimately
            empty - that is a correct result, not a failed arm.
          items:
            type: string
      example:
        query: During INC-2000 on checkout-prod-01, why did checkout writes appear to hang while reads continued? Determine whether CHG-1000 or CHG-1001 caused the incident, identify the customer impact visible to the current principal, explain what evidence rules out the alternative change, and cite the lock evidence and approved runbook supporting both immediate recovery and the preventive follow-up.
        principal:
          scopes:
          - workshop
          principals: []
        filters:
          cluster_id: checkout-prod-01
        controls:
          mode: hybrid
          candidate_pool: 24
          rrf_k: 60
          text_weight: 2.0
          vector_weight: 1.0
          fuzzy_weight: 1.0
          fuzzy_threshold: 0.3
          ef_search: 40
          iterative_scan: strict_order
          rerank: false
        id_tokens:
        - INC-2000
        - CHG-1000
        - CHG-1001
        unresolved_id_tokens: []
...
    TraverseRequest:
      type: object
      required: [seed_evidence_ids, principal, max_depth]
      properties:
        seed_evidence_ids:
          type: array
          items:
            type: string
        principal:
          $ref: '#/components/schemas/Principal'
        max_depth:
          type: integer
          minimum: 0
          maximum: 3
        include_inferred:
          type: boolean
          default: true
...
    TraverseResponse:
      type: object
      description: >
        Relationship paths under ACL. An ACL-invisible node is absent, not a pruned
        branch reported as pruned: the check is applied to the seeds and to both endpoints
        of every hop, so an invisible node cannot appear even as an intermediate.
      required: [contract_version, request_id, nodes, edges]
      properties:
        contract_version:
          type: string
        request_id:
          type: string
        seed_evidence_ids:
          type: array
          description: Seeds as applied, after ACL filtering. A seed the principal cannot see is dropped here.
          items:
            type: string
        max_depth:
          type: integer
        include_inferred:
          type: boolean
        nodes:
          type: array
          description: >
            TYPE CHANGE, additive within contract_version 1.0.0. `nodes` items were
            previously bare evidence_id strings and are now TraverseNode objects carrying
            evidence_id, kind, title, depth and paths. A pre-widening consumer that read
            `nodes[i]` as a string must read `nodes[i].evidence_id`; the string form is no
            longer emitted, and no producer shipped it, which is why this is a description
            change rather than a version bump.
          items:
            $ref: '#/components/schemas/TraverseNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/Edge'
        hidden_by_acl_count:
          type: integer
          description: >
            How many hops were not taken because an endpoint was invisible to the
            principal. Facilitator diagnostics only; the answer must never mention that
            withheld evidence exists.
    CompareRequest:
      type: object
      required: [evidence_ids, principal]
      properties:
        evidence_ids:
          type: array
          items:
            type: string
        principal:
          $ref: '#/components/schemas/Principal'
...
    CompareResponse:
      type: object
      required: [contract_version, request_id, verdicts]
      properties:
        contract_version:
          type: string
        request_id:
          type: string
        verdicts:
          type: array
          items:
            $ref: '#/components/schemas/Verdict'
    ExplainRankingRequest:
      type: object
      required: [run_id, evidence_id]
      properties:
        run_id:
          type: string
        evidence_id:
          type: string
    ExplainRankingResponse:
      type: object
      description: >
        Self-contained ranking explanation for one candidate in one run. Positions, raw
        arm diagnostics, per-arm RRF contribution, the fusion controls that produced the
        score, and the rerank delta are all present, so the score can be recomputed from
        this response alone without re-running retrieval.
      required: [contract_version, request_id, evidence_id, rrf_score, fused_rank]
      properties:
        contract_version:
          type: string
        request_id:
          type: string
        run_id:
          type: string
          description: The persisted run this explanation was read from.
        evidence_id:
          type: string
        text_position:
          type: integer
          nullable: true
        vector_position:
          type: integer
          nullable: true
        fuzzy_position:
          type: integer
          nullable: true
        text_score:
          type: number
          nullable: true
          description: Raw ts_rank from the text arm. Null when the text arm did not return this item.
        vector_distance:
          type: number
          nullable: true
          description: >
            Raw cosine distance from the vector arm - lower is nearer. Null when the vector
            arm did not return this item. It is a distance, not a similarity, and never a
            probability.
        fuzzy_score:
          type: number
          nullable: true
          description: Raw pg_trgm similarity for the probed token. Null when the fuzzy arm did not fire or did not return this item.
        rrf_terms:
          $ref: '#/components/schemas/RrfTerms'
        rrf_score:
          type: number
          description: Sum of rrf_terms. Reproducible from the positions and `controls` on this response.
        controls:
          $ref: '#/components/schemas/SearchControls'
        rerank_score:
          type: number
          nullable: true
        fused_rank:
          type: integer
        final_rank:
          type: integer
          nullable: true
        rerank_delta:
          type: integer
          nullable: true
          description: >
            fused_rank minus final_rank. Positive means the reranker promoted this
            candidate, negative means it demoted it, zero means the rerank agreed with
            fusion. Null when controls.rerank was false, in which case final_rank carries
            no independent information.
        explanation:
          type: string
    SynthesizeRequest:
      type: object
      required: [run_id]
      properties:
        run_id:
          type: string
        replay_only:
          type: boolean
          default: false
...